ApplicationSecurity/Solutions/set2/XSSGoat/requirements.txt
Django==1.10.5
django-csp==3.0
djangorestframework==3.5.4
psycopg2==2.7
